+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 ## [Unreleased]
 
+### Added
+
+- Druid DB connections can now be configured in a custom resource ([#71]).
+
 ### Changed
 
 - Shut down gracefully ([#70]).
 - All dependencies upgraded. The upgrade to operator-rs 0.8.0 does not
   force the credentials secret to be set anymore in the custom resource
   but it is still required ([#82]).
+- - `operator-rs` `0.8.0` → `0.9.0` ([#71])
 
 [#70]: https://github.com/stackabletech/superset-operator/pull/70
+[#71]: https://github.com/stackabletech/superset-operator/pull/71
 [#82]: https://github.com/stackabletech/superset-operator/pull/82
 
 ## [0.2.0] - 2021-12-17